--- a/fe/fe-core/src/main/java/com/starrocks/sql/analyzer/DecimalV3FunctionAnalyzer.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/analyzer/DecimalV3FunctionAnalyzer.java
@@ -126,7 +126,9 @@ public class DecimalV3FunctionAnalyzer {
             return Arrays.stream(argTypes).map(t -> commonType).toArray(Type[]::new);
         }
 
-        if (FunctionSet.ARRAYS_OVERLAP.equalsIgnoreCase(fnName)) {
+        if (FunctionSet.ARRAYS_OVERLAP.equalsIgnoreCase(fnName) ||
+                FunctionSet.ARRAY_CONTAINS_ALL.equalsIgnoreCase(fnName) ||
+                FunctionSet.ARRAY_CONTAINS_SEQ.equalsIgnoreCase(fnName)) {
             Preconditions.checkState(argTypes.length == 2);
             Type[] childTypes = Arrays.stream(argTypes).map(a -> {
                 if (a.isArrayType()) {
@@ -558,7 +560,9 @@ public class DecimalV3FunctionAnalyzer {
                 newFn.setRetType(new ArrayType(triple.returnType));
                 return newFn;
             }
-            case FunctionSet.ARRAYS_OVERLAP: {
+            case FunctionSet.ARRAYS_OVERLAP:
+            case FunctionSet.ARRAY_CONTAINS_ALL:
+            case FunctionSet.ARRAY_CONTAINS_SEQ: {
                 newFn.setArgsType(argumentTypes);
                 return newFn;
             }